requirements.txt
# Core
scrapy>=2.11
scrapy-playwright>=0.0.35
playwright>=1.45

# DB
psycopg[binary,pool]>=3.2
SQLAlchemy>=2.0
python-dotenv>=1.0

# Utils
python-slugify>=8.0